Refactor: Test Controller 분리
